KR-Messanger
This project aims at developing inexpensive virtual reality (VR) text
messages for everyone on a click of a trigger. The 3ds file which is a
part of animation has played a major role as it is the image file which
is been used. This project enables users to know about the use of VR
Cardboard and its functioning.